Crate tree_sitter_sfapex

This crate provides Salesforce Apex, SOQL, SOSL, and Log syntax support for the tree-sitter parsing library.

Typically, you will use the apex::LANGUAGE constant to add this language to a tree-sitter Parser, and then use the parser to parse some code:

use tree_sitter::Parser;

let code = r#"
public class Test123 {
  public Test123(){}
}
"#;
let mut parser = Parser::new();
let language_fn = tree_sitter_sfapex::apex::LANGUAGE;
parser
    .set_language(&language_fn.into())
    .expect("Error loading Apex parser");
let tree = parser.parse(code, None).unwrap();
assert!(!tree.root_node().has_error());

Modules

apex

sflog

soql

sosl